+++ include/ifm3d/camera_nodelet.h
@@ -35,6 +35,7 @@
 #include <ifm3d/Extrinsics.h>
 #include <ifm3d/SoftOff.h>
 #include <ifm3d/SoftOn.h>
+#include <ifm3d/SyncClocks.h>
 #include <ifm3d/Trigger.h>
 
 namespace ifm3d_ros
@@ -61,6 +62,8 @@ namespace ifm3d_ros
     bool Trigger(ifm3d::Trigger::Request& req, ifm3d::Trigger::Response& res);
     bool SoftOff(ifm3d::SoftOff::Request& req, ifm3d::SoftOff::Response& res);
     bool SoftOn(ifm3d::SoftOn::Request& req, ifm3d::SoftOn::Response& res);
+    bool SyncClocks(ifm3d::SyncClocks::Request& req,
+                    ifm3d::SyncClocks::Response& res);
 
     //
     // This is our main publishing loop and its helper functions
@@ -83,6 +86,8 @@ namespace ifm3d_ros
     double soft_on_timeout_tolerance_secs_;
     int soft_off_timeout_millis_;
     double soft_off_timeout_tolerance_secs_;
+    float frame_latency_thresh_;
+    bool sync_clocks_;
 
     std::string frame_id_;
     std::string optical_frame_id_;
@@ -116,6 +121,7 @@ namespace ifm3d_ros
     ros::ServiceServer trigger_srv_;
     ros::ServiceServer soft_off_srv_;
     ros::ServiceServer soft_on_srv_;
+    ros::ServiceServer sync_clocks_srv_;
 
     //
     // We use a ROS one-shot timer to kick off our publishing loop.